requirements.txt
requests
beautifulsoup4
lxml
pandas
playwright
apache-airflow-providers-postgres
psycopg2-binary
openai
python-dotenv
SQLAlchemy

# DBT e dependências
dbt-core>=1.5.0
dbt-postgres>=1.5.0

# Visualização e gráficos
streamlit
plotly
